+++ aplikasi/templates/sapi/sapi_add.html
@@ -15,6 +15,7 @@
 
 {% block scripts %}
   {{ super() }}
+  {{ moment.include_moment() }}
   <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.18.1/moment.min.js"></script>
   <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/bootstrap-datetimepicker/4.17.47/js/bootstrap-datetimepicker.min.js"></script>
   <script type="text/javascript">
