+++ aplikasi/user/user_models.py
@@ -1,0 +1,15 @@
+from .. import db
+class User(db.Model):
+    __tablename__ = 'user'
+    id = db.Column(db.Integer(), primary_key=True)
+    username = db.Column(db.String(255), nullable=False, index=True, unique=True)
+    password = db.Column(db.String(255))
+    last_login = db.Column(db.DateTime, default=datetime.utcnow)
+    created_at = db.Column(db.DateTime, default=db.func.current_timestamp())
+    update_at = db.Column(db.DateTime, default=db.func.current_timestamps(), onupdate=db.func.current_timestamp())
+
+    def __init__(self, username):
+        self.username = username
+    
+    def __repr__(self):
+        return '<user : {}>'.format(self.username)
